// Frontend/src/app/shared/enums/tipo-processoEnum.enum.ts
export enum TipoProcessoEnum {
  civelConsumidor = 1,
  trabalhista = 2,
  administrativo = 3,
  tributarioAdministrativo = 4,
  tributarioJudicial = 5,
  trabalhistaAdministrativo = 6,
  juizadoEspecial = 7,
  expressinho = 8,
  civelEstrategico = 9,
  descumprimentoAdministrativo = 10,
  denunciaFiscal = 11,
  civelAdministrativo = 12,
  criminal = 13,
  criminalAdministrativo = 14,
  criminalJudicial = 15,
  oficioCivelAdministrativo = 16,
  procon = 17,
  PEX = 18,
  servicos = 30

}